requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==2.1.7
django-geojson==2.10.0
django-leaflet==0.26.0
gunicorn==20.0.4
jsonfield==2.0.2
olefile==0.46
Pillow==4.2.1
psycopg2-binary==2.8.5
pytz==2020.1
six==1.10.0
sqlparse==0.3.1
whitenoise==5.0.1
